Finished Project #6, disposed of GUI parts

+++ g-Tests/src/project24/Dec2bin.java
@@ -9,19 +9,19 @@ import java.util.Scanner;
 public class Dec2bin {
 	private static String s = "";
 
-	private static int getRemainder(int n) {
+	private static long getRemainder(long n) {
 		return n % 2;
 	}
 	
 	public static void main(String[] args) {
 		Scanner sc = new Scanner(System.in);
 		System.out.print("Number you want to convert:");
-		int n = sc.nextInt();
+		long n = sc.nextLong();
 		System.out.println("In binary, your number is:" + convertToBinary(n));
 	}
 	
-	public static String convertToBinary(int n) {
-		s = s + Integer.toString(getRemainder(n));
+	public static String convertToBinary(long n) {
+		s = s + Long.toString(getRemainder(n));
 		if(n != 0) {
 			convertToBinary((int)Math.floor(n / 2));
 		}
